% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullDesign.R
\name{getFullDesign}
\alias{getFullDesign}
\title{Generating the Full Design Matrix}
\usage{
getFullDesign(formula, data, center = TRUE)
}
\arguments{
\item{formula}{an object of class "\code{\link[=formula]{formula()}}" (or one that can be coerced to that class); a symbolic description of the model to be fitted.}

\item{data}{a data frame (or object coercible by \code{as.data.frame} to a data frame) containing the variables in the model.}

\item{center}{Boolean, whether to center continuous predictors. \code{TRUE} by default for \code{lmabc}.}
}
\description{
\code{getFullDesign} is used to generate the design matrix for linear regression with ABCs.
}
\section{Details}{
The full design matrix is similar to the design matrix in standard linear regression. However, instead of removing one level from each categorical variable, \code{getFullDesign} retains all levels by manipulating contrast matrices.

The full design matrix is used with the constraint matrix from \code{\link[=getConstraints]{getConstraints()}} to fit a model with the ABCs.

This method is called by \code{\link[=lmabc]{lmabc()}}. This method is useful for implementing additional variations of the ABCs.
}

\section{Value}{
\code{getFullDesign} returns a matrix representing the full design matrix for a particular regression model.
}

\examples{
# full design matrix
getFullDesign(Sepal.Length ~ Petal.Length + Species + Petal.Length*Species, data = iris)

# lm's default design matrix, which removes a reference category
model.matrix(lm(Sepal.Length ~ Petal.Length + Species + Petal.Length*Species, data = iris))

}
\seealso{
\code{\link[=lmabc]{lmabc()}} for a use case of \code{getFullDesign}.
}
