% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmabc.R
\name{glmabc}
\alias{glmabc}
\title{Fitting Generalized Linear Models with Abundance-Based Constraints (ABCs)}
\usage{
glmabc(formula, family = stats::gaussian, data, ..., cprobs = NULL)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link[stats]{family}} for details of
    family functions.)}

\item{data}{a data frame (or object coercible by \code{as.data.frame} to a data frame) containing the variables in the model.}

\item{...}{
    For \code{glm}: arguments to be used to form the default
    \code{control} argument if it is not supplied directly.

    For \code{weights}: further arguments passed to or from other methods.
  }

\item{cprobs}{a named list with an entry for each named categorical variable in the model, specifying the probabilities of each category.}
}
\description{
\code{glmabc} is used to fit generalized linear models using abundance-based constraints (ABCs).
Like \code{\link[stats:glm]{stats::glm}}, the model is specified by giving a symbolic description of the linear predictor
and a description of the error distribution. ABCs provide more equitable and interpretable output
for regression models with categorical covariates.
}
\section{Details}{
A \code{glmabc} model is specified identically to the corresponding \code{glm} model.
At this time, \code{glmabc} only supports a single response variable, and the data must be passed into the \code{data} parameter.
}

\section{Differences from \code{glm}}{
The default approach for generalized linear regression with categorical covariates
is reference group encoding (RGE): one category is selected as the "reference group"
for each categorical variable, and all results are presented
relative to this group. This output is inequitable and can be
misleading, especially for categorical covariates like race/ethnicity,
gender identity, religion, national origin, etc.

For example, suppose an analyst fits a generalized linear model of the form \code{y ~ x + race + x:race},
where \code{x} is a continuous covariate. This model allows for race-specific
coefficients on \eqn{x}. However, RGE requires a reference group, typically
White (or selected alphabetically), and the \code{glm} output is presented relative to this group:
\code{(Intercept)} refers to the intercept for the White group and
\code{x} refers to the coefficients on \eqn{x} for the White group. The remaining
race-specific parameters are all presented as differences relative to the White group.
Clearly, this output is inequitable: it elevates one group (White) above others,
and presents this group as "normal" while others are "deviations from normal".
The output is also unclear: that the \code{(Intercept)} and \code{x} effects
refer \emph{only} to the reference (White) group is nowhere mentioned in any default
output.

\code{glmabc} addresses these issues. ABCs parametrize the regression model so that the main effect terms, here
\code{(Intercept)} and \code{x}, are \emph{averages} of the race-specific terms.
For instance, the coefficients on \eqn{x} represent the race-averaged effect
of \eqn{x}, and the race-specific coefficients on \eqn{x} represent deviations
from the average, rather than from the reference (White) group. The
notion of "average" derives from the argument \code{cprob}: these can be
input by the user (e.g., population proportions), otherwise
they will be set to the sample proportions for each group.
}

\section{Value}{
\code{glmabc} returns an object of classes "glmabc" and "lmabc."
Many generics commonly used for \code{glm} objects have been implemented for \code{glmabc}:
\code{summary}, \code{coefficients}, \code{plot}, \code{predict}, and more.
See the DESCRIPTION file for all implemented S3 methods.
}

\examples{
mtcars$cyl <- as.factor(mtcars$cyl)
fit <- glmabc(am ~ mpg + cyl + mpg:cyl, family = "binomial", data = mtcars)
summary(fit)

predict(fit, newdata = data.frame(mpg = 21, cyl = "6"), type = 'response')

}
\seealso{
\code{\link[stats:glm]{stats::glm()}} for the standard generalized linear regression implementation in R.
}
