% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lmabc}
\alias{summary.lmabc}
\title{Summarizing Linear ABC Model Fits}
\usage{
\method{summary}{lmabc}(object, ...)
}
\arguments{
\item{object}{An object of class "lmabc", usually, a direct result of a call to lmabc}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of summary statistics of the fitted linear ABC
model given in object.
}
\description{
\code{summary.lmbc} summary method for class "lmabc".
}
\details{
This function uses the summary.lm functionality. To learn more about the original
functionality go to \code{\link[stats]{summary.lm}}.
}
