% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm_abc}
\alias{lm_abc}
\title{\code{lm_abc} is used to fit linear models using abundance-based constraints.}
\usage{
lm_abc(formula, data, ..., cprobs = NULL)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{lm} is called.}

\item{...}{additional arguments to be passed to the low level
    regression fitting functions (see below).}

\item{cprobs}{a named list with an entry for each named categorical variable in the model, specifying the probabilities of each category}
}
\description{
\code{lm_abc} is used to fit linear models using abundance-based constraints.
}
