% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmabc-package.R
\docType{package}
\name{lmabc-package}
\alias{lmabc-package}
\alias{_PACKAGE}
\title{lmabc: Linear Regression with Abudance Based Constraints}
\description{
lmabc introduces an implementation of linear regression with abundance-based constraints, which seeks to increase equity and interpretability in statistical analyses. Traditional linear regression models with categorical predictors choose a "baseline" category, generally the most prevalent one, which tends to be non-Hispanic White for race. This creates inequities in the presentation and interpretation of such models. lmabc uses an averaging approach to change the baseline from a single arbitrarily-chosen level to the global average.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://prayaggordy.github.io/lmabc/}
  \item \url{https://github.com/prayaggordy/lmabc}
}

}
\author{
\strong{Maintainer}: Dan Kowal \email{daniel.r.kowal@gmail.com}

Authors:
\itemize{
  \item Prayag Gordy
  \item Virginia Baskin
  \item Jai Uparkar
}

Other contributors:
\itemize{
  \item Caleb Fikes [contributor]
}

}
\keyword{internal}
