% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmabc-package.R
\docType{package}
\name{lmabc-package}
\alias{lmabc-package}
\alias{_PACKAGE}
\title{lmabc: Linear Regression with Abundance-Based Constraints (ABCs)}
\description{
lmabc provides estimation and inference for linear regression with categorical covariates (race, sex, etc.). Common strategies, including the defaults in lm, select a "reference" group (e.g,. White, Male, etc.) and present all results relative to this group. However, this approach suffers from alarming biases, statistical inefficiencies, and limited interpretability. lmabc uses abundance-based constraints (ABCs) that (1) fully eliminate these biases, (2) provide appealing statistical properties related to estimation invariance and improved statistical power, and (3) produce more interpretable model output. lmabc reproduces the core functionality of lm and includes functions for generalized (logistic, Poisson, etc.) and penalized (lasso, ridge, etc.) linear models. Details are provided in Kowal (2024) \doi{10.1101/2024.01.04.23300033}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://prayaggordy.github.io/lmabc/}
  \item \url{https://github.com/prayaggordy/lmabc}
}

}
\author{
\strong{Maintainer}: Dan Kowal \email{daniel.r.kowal@gmail.com}

Authors:
\itemize{
  \item Prayag Gordy
  \item Virginia Baskin
  \item Jai Uparkar
}

Other contributors:
\itemize{
  \item Caleb Fikes [contributor]
}

}
\keyword{internal}
