% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmabc.R
\name{lmabc}
\alias{lmabc}
\title{Fitting Linear Models with Abundance-Based Constraints (ABCs)}
\usage{
lmabc(formula, data, ..., cprobs = NULL)
}
\arguments{
\item{formula}{an object of class "\code{\link[=formula]{formula()}}" (or one that can be coerced to that class); a symbolic description of the model to be fitted.}

\item{data}{a data frame (or object coercible by \code{as.data.frame} to a data frame) containing the variables in the model.}

\item{...}{For \code{lm()}: additional arguments to be passed to the low level
    regression fitting functions (see below).}

\item{cprobs}{an optional named list with an entry for each named categorical variable in the model, specifying the probabilities of each category, which must sum to 1. By default, \code{cprobs} will be calculated from the proportions in the data.}
}
\description{
\code{lmabc} is used to fit linear models using abundance-based constraints (ABCs).
ABCs provide more equitable and interpretable output
for regression models with categorical covariates.
}
\section{Details}{
An \code{lmabc} model is specified identically to the corresponding \code{lm} model. At this time, \code{lmabc} only supports a single response variable, and the data must be passed into the \code{data} parameter.
}

\section{Differences from \code{lm}}{
The default approach for linear regression with categorical covariates
is reference group encoding (RGE): one category is selected as the "reference group"
for each categorical variable, and all results are presented
relative to this group. This output is inequitable and can be
misleading, especially for categorical covariates like race/ethnicity,
gender identity, religion, national origin, etc.

For example, suppose an analyst fits a model of the form \code{y ~ x + race + x:race},
where \code{x} is a continuous covariate. This model allows for race-specific
coefficients on \eqn{x}. However, RGE requires a reference group, typically
White (or selected alphabetically), and the \code{lm} output is presented relative to this group:
\code{(Intercept)} refers to the intercept for the White group and
\code{x} refers to the coefficients on \eqn{x} for the White group. The remaining
race-specific parameters are all presented as differences relative to the White group.
Clearly, this output is inequitable: it elevates one group (White) above others,
and presents this group as "normal" while others are "deviations from normal".
The output is also unclear: that the \code{(Intercept)} and \code{x} effects
refer \emph{only} to the reference (White) group is nowhere mentioned in any default
output.

\code{lmabc} addresses these issues. ABCs parametrize the regression model so that the main effect terms, here
\code{(Intercept)} and \code{x}, are \emph{averages} of the race-specific terms.
For instance, the coefficients on \eqn{x} represent the race-averaged effect
of \eqn{x}, and the race-specific coefficients on \eqn{x} represent deviations
from the average, rather than from the reference (White) group. The
notion of "average" derives from the argument \code{cprob}: these can be
input by the user (e.g., population proportions), otherwise
they will be set to the sample proportions for each group.
}

\section{Value}{
\code{lmabc} returns an object of class "lmabc." Many generics commonly used for \code{lm} objects have been implemented for \code{lmabc}: \code{summary}, \code{coefficients}, \code{plot}, \code{predict}, and more. See the DESCRIPTION file for all implemented S3 methods.
}

\examples{
fit <- lmabc(Sepal.Length ~ Petal.Length + Species + Petal.Length*Species, data = iris)
summary(fit)

predict(fit, newdata = data.frame(Petal.Length = 1.5, Species = "setosa"))

}
\seealso{
\code{\link[=lm]{lm()}} for the standard linear regression implementation in R.
}
