% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lmabc}
\alias{summary.lmabc}
\title{Summarizing Linear ABC Model Fits}
\usage{
\method{summary}{lmabc}(object, ...)
}
\arguments{
\item{object}{An object of class "lmabc", usually, a direct result of a call to \link{lm_abc}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of summary statistics of the fitted linear ABC model given in object.
}
\description{
\code{summary} method for class "lmabc".
}
\seealso{
\link[stats:summary]{stats::summary}
}
